requirements.txt
drf-spectacular>=0.24.2,<0.25
Django>=4.1.1,<4.2
djangorestframework>=3.13.1,<3.14
django-cors-headers==3.12.0
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.24.2,<0.25
python-keycloak>=2.5.0,<2.6
PyJWT>=2.5.0

# Debugging
django-cors-headers==3.12.0
flake8>=5.0.4,<5.1
flake8-bandit>=4.1.1
flake8-docstrings>=1.6.0

